[WIP] Add hello-json example
This example uses `ppx_deriving_yojson` to generate the `data_of_yojson` json
parsing function which returns an OCaml datatype matching the type `data`, which
is defined be a record with two fields, etc. This appears to work, since
`merlin` reports no errors when checking the code using the generated `.merlin`
file. However, `jbuilder build` fails because of it's [limited support for ppx
drivers][noppx].

TODO:
- fix build, or wait for upstream jbuilder+ppx_deriving fix

[noppx]: https://github.com/mirage/mirage/issues/818#issuecomment-304672837